refactor(balances): move checkUsageAlerts into usageAlerts/check
